stt-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
setuptools>=65.0.0
numpy<2.0.0
openai-whisper
python-multipart==0.0.6
torch==2.1.2
